Split guest cloud upgrade tests
